Consolidate module entrypoint, add keywords
